Basic YAML support & better error handling.
