+++ Cargo.bazel.lock
@@ -1,5 +1,5 @@
 {
-  "checksum": "a7d54e3a0880043148cbff51b4a002dea0313eb750503c45f549ab11a240ef53",
+  "checksum": "490c7fe5080d3e395c6d807c8075456195c6bc29931181e804a7428e9a4cfa82",
   "crates": {
     "addr2line 0.19.0": {
       "name": "addr2line",
@@ -1482,13 +1482,13 @@
       },
       "license": "Apache-2.0"
     },
-    "console 0.15.3": {
+    "console 0.15.4": {
       "name": "console",
-      "version": "0.15.3",
+      "version": "0.15.4",
       "repository": {
         "Http": {
-          "url": "https://crates.io/api/v1/crates/console/0.15.3/download",
-          "sha256": "5556015fe3aad8b968e5d4124980fbe2f6aaee7aeec6b749de1faaa2ca5d0a4c"
+          "url": "https://crates.io/api/v1/crates/console/0.15.4/download",
+          "sha256": "c9b6515d269224923b26b5febea2ed42b2d5f2ce37284a4dd670fedd6cb8347a"
         }
       },
       "targets": [
@@ -1544,7 +1544,7 @@
           }
         },
         "edition": "2018",
-        "version": "0.15.3"
+        "version": "0.15.4"
       },
       "license": "MIT"
     },
@@ -2874,7 +2874,7 @@
         "deps": {
           "common": [
             {
-              "id": "console 0.15.3",
+              "id": "console 0.15.4",
               "target": "console"
             },
             {
@@ -11437,6 +11437,64 @@
       },
       "license": "MIT OR Apache-2.0"
     },
+    "serde_yaml 0.9.16": {
+      "name": "serde_yaml",
+      "version": "0.9.16",
+      "repository": {
+        "Http": {
+          "url": "https://crates.io/api/v1/crates/serde_yaml/0.9.16/download",
+          "sha256": "92b5b431e8907b50339b51223b97d102db8d987ced36f6e4d03621db9316c834"
+        }
+      },
+      "targets": [
+        {
+          "Library": {
+            "crate_name": "serde_yaml",
+            "crate_root": "src/lib.rs",
+            "srcs": {
+              "include": [
+                "**/*.rs"
+              ],
+              "exclude": []
+            }
+          }
+        }
+      ],
+      "library_target_name": "serde_yaml",
+      "common_attrs": {
+        "compile_data_glob": [
+          "**"
+        ],
+        "deps": {
+          "common": [
+            {
+              "id": "indexmap 1.9.2",
+              "target": "indexmap"
+            },
+            {
+              "id": "itoa 1.0.5",
+              "target": "itoa"
+            },
+            {
+              "id": "ryu 1.0.12",
+              "target": "ryu"
+            },
+            {
+              "id": "serde 1.0.152",
+              "target": "serde"
+            },
+            {
+              "id": "unsafe-libyaml 0.2.5",
+              "target": "unsafe_libyaml"
+            }
+          ],
+          "selects": {}
+        },
+        "edition": "2021",
+        "version": "0.9.16"
+      },
+      "license": "MIT OR Apache-2.0"
+    },
     "siphasher 0.2.3": {
       "name": "siphasher",
       "version": "0.2.3",
@@ -13725,6 +13783,63 @@
       },
       "license": "MIT OR Apache-2.0"
     },
+    "unsafe-libyaml 0.2.5": {
+      "name": "unsafe-libyaml",
+      "version": "0.2.5",
+      "repository": {
+        "Http": {
+          "url": "https://crates.io/api/v1/crates/unsafe-libyaml/0.2.5/download",
+          "sha256": "bc7ed8ba44ca06be78ea1ad2c3682a43349126c8818054231ee6f4748012aed2"
+        }
+      },
+      "targets": [
+        {
+          "Library": {
+            "crate_name": "unsafe_libyaml",
+            "crate_root": "src/lib.rs",
+            "srcs": {
+              "include": [
+                "**/*.rs"
+              ],
+              "exclude": []
+            }
+          }
+        },
+        {
+          "Binary": {
+            "crate_name": "run-parser-test-suite",
+            "crate_root": "src/bin/run-parser-test-suite.rs",
+            "srcs": {
+              "include": [
+                "**/*.rs"
+              ],
+              "exclude": []
+            }
+          }
+        },
+        {
+          "Binary": {
+            "crate_name": "run-emitter-test-suite",
+            "crate_root": "src/bin/run-emitter-test-suite.rs",
+            "srcs": {
+              "include": [
+                "**/*.rs"
+              ],
+              "exclude": []
+            }
+          }
+        }
+      ],
+      "library_target_name": "unsafe_libyaml",
+      "common_attrs": {
+        "compile_data_glob": [
+          "**"
+        ],
+        "edition": "2021",
+        "version": "0.2.5"
+      },
+      "license": "MIT"
+    },
     "url 1.7.2": {
       "name": "url",
       "version": "1.7.2",
@@ -14998,6 +15113,10 @@
             {
               "id": "serde_json 1.0.91",
               "target": "serde_json"
+            },
+            {
+              "id": "serde_yaml 0.9.16",
+              "target": "serde_yaml"
             },
             {
               "id": "slug 0.1.4",
@@ -16781,6 +16900,7 @@
   "binary_crates": [
     "cc 1.0.78",
     "pulldown-cmark 0.9.2",
+    "unsafe-libyaml 0.2.5",
     "wasm-pack 0.10.3"
   ],
   "workspace_members": {
