+++ Cargo.toml
@@ -19,6 +19,7 @@ lazy_static = "1.4"
 pulldown-cmark = "0.9"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
+serde_yaml = "0.9"
 slug = "0.1"
 wasm-bindgen = "0.2"
 wasm-bindgen-futures = "0.4"
